+++ ledger-core/drafts/unused/pgbuf_flushed.go
@@ -3,7 +3,7 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
 
-package dropstorage
+package unused
 
 import (
 	"io"
